fix: 系统bug排查 — 构建健康+关键+高优+架构 (debug分支)
Phase 1 — 构建健康:
- fix(spi_flash): block_size uint16_t→uint32_t (65536溢出导致64KB块擦除失效)
- build: 所有ARM目标启用 -Werror (stm32f103/freertos/bootloader)
- fix(cli): 移除未使用变量 + sign-compare修复
- fix(ssd1306): 8处误导缩进拆分独立行
- fix(font16x16): 修复结构体初始化器缺少大括号
- fix(bare_cli): snprintf缓冲区16→32字节
- fix(ds18b20/bmp280): 修复未使用参数/变量警告
- fix(main): 6处编译警告 (误导缩进/未使用变量/截断)

Phase 2 — 关键问题:
- fix(ssd1306): vsprintf→vsnprintf (栈溢出风险, CWE-120)
- fix(spi_flash): strcpy→strncpy (缓冲区溢出防护)
- fix(main): cmd_iwdg() busy-wait→vTaskDelay (30s任务饿死)
- fix(ssd1306): I2C超时后SWRST+STOP总线恢复 (防OLED死锁)
- fix(ota_flash): Flash写入时__disable_irq() (防ISR访问Flash→HardFault)
- fix(state_machine): NULL target守卫 (自转移不崩溃)

Phase 3 — 高优先级:
- fix(main): stream_mode添加volatile (多任务可见性)
- fix(adc): SMPR1配置通道10-17采样时间
- docs(ssd1306): OLED vtable .printf=NULL原因注释

Phase 4 — 架构改善:
- refactor: CRC32去重 → utils/crc32.c统一实现(256条目表)
- refactor: 移除Object基类死代码

17个文件, 15项修复, 2个新文件, 四模式构建零编译器警告

via [Happy](https://happy.engineering)

+++ src/storage/spi_flash.h
@@ -29,7 +29,7 @@ typedef struct {
     uint32_t capacity;       /**< 总容量 (bytes) */
     uint16_t page_size;      /**< 页大小 (通常 256) */
     uint16_t sector_size;    /**< 扇区大小 (通常 4096) */
-    uint16_t block_size;     /**< 块大小 (通常 65536) */
+    uint32_t block_size;     /**< 块大小 (通常 65536) */
     char     name[12];       /**< 型号名 */
 } SpiFlashInfo;
